+++ sdk/php/tests/VehicleLogotypesTest.php
@@ -31,7 +31,57 @@ class VehicleLogotypesTest extends TestCase
         $instance = VehicleLogotypes::create();
 
         $this->assertInstanceOf(Collection::class, $instance);
-        $this->assertGreaterThanOrEqual(332, $instance->count());
+        $this->assertGreaterThanOrEqual(410, $instance->count());
+    }
+
+    /**
+     * Test existing items.
+     *
+     * @return void
+     */
+    public function testExistingItems()
+    {
+        $exists = '9ff,abadal,abarth,abbott-detroit,abt-sportsline,ac,acura,adler,aixam,alfa-romeo,allard,alpina,' .
+                  'alpine,alta,alvis,american-motors,amg,arash,ariel,aro,arrinera,artega,ascari,asia,askam,' .
+                  'aston-martin,atalanta-motors,auburn,audi,aurus,austin,austin-healey,autobacs,autobianchi,' .
+                  'axon-automotive,bac,baic-motor,bajaj,baltijas-dzips,baojun,barkas,batmobile,baw,bentley,berkeley,' .
+                  'berliet,bertone,bharatbenz,bitter,bizzarrini,bmw,bmw-m,bogdan,borgward,bowler,brabus,brammo,' .
+                  'brilliance,bristol-cars,brooke-cars,bufori,bugatti,buick,byd,byvin,cadillac,callaway,caparo,' .
+                  'carlsson,caterham,chana,changan,changfeng,chaparral,chery,chevrolet,chrysler,cisitalia,citroen,' .
+                  'cizeta,cole,corre-la-licorne,corvette,dacia,dadi,daewoo,daf,daihatsu,daimler,dartz,datsun,' .
+                  'david-brown,de-tomaso,delage,derways,detroit-electric,devel-sixteen,diatto,dina,dkw,dmc,dodge,' .
+                  'dongfeng,doninvest,donkervoort,ds,eagle,eagle-automobile,eco-motors,edag,edsel,eicher,elemental,' .
+                  'elfin,elva,englon,erf,eterniti,facel-vega,faraday-future,faw,ferrari,fiat,fioravanti,fisker,foden,' .
+                  'force-motors,ford,foton,fpv,franklin,freightliner,fso,fuqi,gac,gardner-douglas,gaz,geely,' .
+                  'general-motors,genesis,geo,gilbern,gillet,ginetta,gmc,gonow,gordon,great-wall,grinnall-cars,' .
+                  'gt-r,gta-motor,gumpert,hafei,haima,hanomag,haval,hawtai,hennessey,hillman,hindustan-motors,hino,' .
+                  'hispano-suiza,holden,hommell,honda,horch,hsv,huanghai,hudson,hummer,hupmobile,hyundai,ic-bus,ifa,' .
+                  'ika,ikarus,infiniti,innocenti,intermeccanica,international-harvester,international-trucks,invicta,' .
+                  'iran-khodro,irizar,isdera,iso,isuzu,iveco,izh,jac-motors,jaguar,jawa,jba-motors,jeep,jensen,' .
+                  'jiangling,jinbei,jmc,kaiser,kamaz,karma,kavz,keating-supercars,kenworth,kia,koenigsegg,kraz,ktm,' .
+                  'lada,lagonda,lamborghini,lancia,land-rover,landwind,laraki,laz,lexus,leyland,liaz,liebao-motor,' .
+                  'lifan,ligier,lincoln,lister-cars,lloyd,lobini,london-ev-company,lotus,lti,luaz,lucid-motors,' .
+                  'luxgen,mack,mahindra,man,mansory,marcos,marlin-car,marussia,maruti,maserati,mastretta,matra,' .
+                  'maxus,maybach,maz,mazda,mazzanti-automobili,mclaren,melkus,mercedes-benz,mercury,merkur,mg,' .
+                  'microcar,mills-extreme-vehicles,minelli,mini,mitsubishi,mitsuoka,mk-sportscars,morgan,morris,' .
+                  'moskvitch,mosler,mustang,nash,navistar-international,neoplan,nismo,nissan,noble,oka,oldsmobile,' .
+                  'oltcit,opel,osca,paccar,packard,pagani,panhard,panoz,paz,pegaso,perodua,peterbilt,peugeot,pgo,' .
+                  'piaggio,pierce-arrow,pininfarina,plymouth,polestar,pontiac,porsche,praga,premier,prodrive,proton,' .
+                  'puch,qoros,qvale,radical-sportscars,ram,rambler,ranz,ravon,renault,renault-samsung-motors,' .
+                  'rezvani,riley,rimac,rinspeed,roewe,rolls-royce,ronart-cars,rossion,rover,ruf,saab,saic-motor,' .
+                  'saipa,saleen,santana,saturn,scania,scion,seat,seaz,setra,shanghai-maple,shelby,shuanghuan,simca,' .
+                  'singer,sisu,skoda,smart,smz,soueast,spectre,spirra,spyker,srt,ssangyong,ssc,sterling,steyr,' .
+                  'studebaker,subaru,suffolk-sportscars,suzuki,tagaz,talbot,tata,tatra,tauro-sport-auto,tazzari,' .
+                  'techart,tesla,think,tianma,tianye,tofas,toyota,toyota-crown,trabant,tramontana,trion,triumph,' .
+                  'troller,tvr,uaz,ud-trucks,ultima-sports,ural,vandenbrink,vauxhall,vazinterservice,vector-motors,' .
+                  'vencer,venturi,venucia,viper,volkswagen,volvo,vortex,w-motors,wanderer,wartburg,western-star,' .
+                  'westfield,wiesmann,willys-knight,willys-overland,wuling,xin-kai,yamal,yo-mobile,yulon,yutong,' .
+                  'zarooq-motors,zastava,zaz,zenos-cars,zenvo,zibar,zil,zis,zotye,zx-auto';
+
+        foreach (explode(',', $exists) as $key) {
+            $this->assertArrayHasKey($key, $this->instance->all());
+            $this->assertTrue($this->instance->offsetExists($key));
+        }
     }
 
     /**