+++ CHANGELOG.md
@@ -4,13 +4,13 @@
 
 ### Added
 
-- New logotypes
+- New logotypes: `adler, chaparral, ika, marussia, santana, ural, allard, dadi, ikarus, maruti, seaz, vazinterservice, asia, derways, invicta, matra, shanghai-maple, vortex, aurus, izh, minelli, shuanghuan, willys-knight, austin-healey, doninvest, jinbei, moskvitch, smz, xin-kai, bajaj, eagle, jmc, nash, spectre, yamal, baltijas-dzips, eco-motors, kavz, neoplan, steyr, yo-mobile, barkas, fso, kraz, oka, tagaz, yutong, batmobile, fuqi, packard, tazzari, baw, laz, paz, think, zibar, bogdan, gordon, liaz, piaggio, tianma, zil, byvin, hanomag, liebao-motor, puch, tianye, zis, callaway, huanghai, lti, qvale, tofas, zx-auto, chana, ifa, luaz, ravon, trabant`
 
 ## v1.1.0
 
 ### Added
 
-- Images uses CDN proxy service imgix.com
+- Images uses CDN proxy service `imgix.com`
 - PHP-sdk
 
 ## v1.0.0